Moved an interface out of public API, as it wasn't supposed to be there.
