+++ subprojects/language-native/src/main/java/org/gradle/language/cpp/internal/DefaultCppApplication.java
@@ -24,7 +24,7 @@ import org.gradle.api.provider.Property;
 import org.gradle.language.cpp.CppApplication;
 import org.gradle.language.cpp.CppExecutable;
 import org.gradle.language.cpp.CppPlatform;
-import org.gradle.language.nativeplatform.PublicationAwareComponent;
+import org.gradle.language.nativeplatform.internal.PublicationAwareComponent;
 import org.gradle.nativeplatform.toolchain.internal.NativeToolChainInternal;
 import org.gradle.nativeplatform.toolchain.internal.PlatformToolProvider;
