+++ subprojects/language-native/src/test/groovy/org/gradle/language/plugins/NativeBasePluginTest.groovy
@@ -34,7 +34,7 @@ import org.gradle.language.ComponentWithBinaries
 import org.gradle.language.ComponentWithOutputs
 import org.gradle.language.ProductionComponent
 import org.gradle.language.internal.DefaultBinaryCollection
-import org.gradle.language.nativeplatform.PublicationAwareComponent
+import org.gradle.language.nativeplatform.internal.PublicationAwareComponent
 import org.gradle.language.nativeplatform.internal.ConfigurableComponentWithExecutable
 import org.gradle.language.nativeplatform.internal.ConfigurableComponentWithLinkUsage
 import org.gradle.language.nativeplatform.internal.ConfigurableComponentWithRuntimeUsage
